OF-999: Allow configuration of BOSH worker thread pool The thread pool that performs processing of data received through the HTTP-based connections of BOSH (as provided by Jetty) should not have a hardcoded upper bound. Instead, the pool size should be configurable.
For other connection types, this is managed by properties named as such:
 - xmpp.client.processing.threads (client sockets)
 - xmpp.multiplex.processing.threads (connection managers)
 - xmpp.component.processing.threads (components)

BOSH does have a similar property: httpbind.client.processing.threads
It's value however is not used to configure the amount of worker threads. It is
used to configure the amount of (NIO) acceptors instead. In NIO, acceptor
threads handle network IO, and hand off processing to the worker threads.

There's little benefit in being able to configure the amount of acceptors, as
Jetty will default to a platform-specific default (based on the number of CPUs
that are available). There is greater benefit in being able to configure the
amount of worker threads (as that's more likely to affect throughput).